2/2 Refactor dirChannels/Inputs/Outputs. Fix use of 'unspecified' directions in the case of using *newFromMatrix.
(Part of previous commit)

There was an inconsistency between how Foa matrix classes used
dirChannels (as a getter for dirInputs/Outputs based on the class) and
how Hoa matrix classes (and FoaKernel classes) used it (dirChannels as
an instance variable, which is returned by dirInputs/Outputs getter
methods, depending on the class). This change make it consistent across
all of the ATK.

Also fixes a bug when assigning ‘set’ attribute to HOA1 matrix.

Also contains various whitespace cleanup and minor style adjustments.