[docs] Add clarifying comments regarding angular frequency vs. spatial frequency.
Spatial frequency, as defined in the wikipedia article linked in the docs, is cycles/m, whereas wavenumber in the ATK follows the physics convention of radians/meter. S o it's related, but is not identical to a spatial frequency. So the added comments disambiguate this, making it more immediately clear to the user what the units of WaveNumber are.